Added safeguard while sending Chat
While the user waits for a reply from the bot, they can type out a message but not send the message. This is done because sending another message while the bot is replying was resulting in having the same ID for the message in firebase and crashed the funtionality inside the application.